pyproject.toml
[project]
name = "research-project"
version = "0.1.0"
description = "Research project template"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "pytest>=7.3.0",
    "python-dotenv>=1.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "datasets>=2.14.0",
    "tokenizers>=0.13.0",
    "accelerate>=0.20.0",
    "scikit-learn>=1.3.0",
    "scikit-image>=0.21.0",
    "fastapi[all]>=0.100.0",
    "requests>=2.31.0",
    "wandb>=0.15.0",
    "h5py>=3.9.0",
    "einops>=0.6.0",
    "dspy-ai>=3.0.0",
    "jsonlines>=3.1.0",
    "peft>=0.17.1",
    "sglang[all]>=0.4.0",
    "safetensors>=0.6.2",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.23.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
